Fix editing view when clicking Create
I entered the root path and immediatly clicked the Create-button, but the editing view was not toggled. It complained about the 'edit' route. What was intended in routes/post.js is instead (also) done in routes/posts.js with this patch. Now it works! I don't know if it's the best way but I guess it is a way. :-) Have a great day, and thanks for the example code!